requirements.txt
alabaster==0.7.12
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
Babel==2.8.0
beautifulsoup4==4.9.3
Brlapi==0.6.6
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.4.3
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.1.2
djangorestframework==3.12.1
docopt==0.6.2
docutils==0.16
feedparser==5.2.1
gitsome==0.8.4
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
imagesize==1.2.0
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
numpydoc==0.9.2
oauth==1.0.1
olefile==0.45.1
packaging==20.3
pexpect==4.2.1
Pillow==5.1.0
pipenv==2020.8.13
ply==3.11
prompt-toolkit==2.0.10
protobuf==3.0.0
psycopg2==2.7.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.7.1
pytz==2016.10
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
uritemplate.py==3.0.2
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
virtualenv-clone==0.5.4
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5
xkit==0.0.0
zope.interface==4.3.2
